scripts/alert/requirements.txt
azure-common==1.1.28
azure-core==1.35.0
azure-storage-common==2.1.0
azure-storage-file==2.1.0